Create Odd Even Linked List
Given a singly linked list, group all odd nodes together followed by
the even nodes. Please note here we are talking about the node number
and not the value in the nodes. You should try to do it in place. The
program should run in O(1) space complexity and O(nodes) time
complexity. The relative order inside both the even and odd groups
should remain as it was in the input.  The first node is considered
odd, the second node even and so on ...